Gperf fixes

+++ include/core_def.h
@@ -244,6 +244,6 @@ static inline String *get_string(Node *node) {
 
 /* prim_op.gperf */
 const struct Primitive *primitive_lookup(register const char *str,
-                                         register size_t len);
+                                         register unsigned int len);
 
 #endif
